Feat: Implement modular demo with switchable graph pages
Redesigned the demo to support multiple, independent graph examples (pages).
Each page showcases specific features of SpaceGraphJS and includes an explanation in the HUD.

Key changes:
- Created `src/examples/pages.js` to define page configurations (ID, title, description, graph creation function).
- Modified `index.html` to include a page selector UI and a HUD display area.
- Updated the main script in `index.html` to dynamically populate the page selector and load the selected page's graph and description.
- Refactored the original demo graph into the first page ('all-features').
- Added two new example pages:
    - 'Node Types Showcase': Demonstrates various node types.
    - 'Edge Types Showcase': Demonstrates various edge styles and properties.
- Added CSS for the new UI elements.
- Deleted the old `src/examples/exampleGraph.js` file as its functionality is now part of `pages.js`.
