+++ index.html
@@ -12,6 +12,13 @@
     <body>
         <div id="toolbar"></div>
 
+        <div id="page-selector">
+            <!-- Buttons will be added here by JavaScript -->
+        </div>
+        <div id="hud">
+            <!-- Page description will be loaded here -->
+        </div>
+
         <div id="mindmap-container">
             <canvas id="webgl-canvas"></canvas>
             <div id="css3d-container"></div>
@@ -36,44 +43,110 @@
         </script>
         <script type="module">
             import * as S from './src/index.js';
-            import { createExampleGraph } from './src/examples/exampleGraph.js';
+            import { pages } from './src/examples/pages.js';
+
+            let space; // Make space globally accessible within this script
+            const pageSelectorEl = S.$('#page-selector');
+            const hudEl = S.$('#hud');
+
+            function populatePageSelector() {
+                if (!pageSelectorEl) return;
+                pageSelectorEl.innerHTML = ''; // Clear existing buttons
+
+                pages.forEach(page => {
+                    const button = document.createElement('button');
+                    button.textContent = page.title;
+                    button.dataset.pageId = page.id;
+                    button.addEventListener('click', () => loadPage(page.id));
+                    pageSelectorEl.appendChild(button);
+                });
+            }
+
+            function updateActiveButton(pageId) {
+                if (!pageSelectorEl) return;
+                const buttons = pageSelectorEl.querySelectorAll('button');
+                buttons.forEach(button => {
+                    if (button.dataset.pageId === pageId) {
+                        button.classList.add('active');
+                    } else {
+                        button.classList.remove('active');
+                    }
+                });
+            }
+
+            async function loadPage(pageId) {
+                if (!space) {
+                    console.error('SpaceGraph not initialized yet.');
+                    return;
+                }
+
+                const page = pages.find(p => p.id === pageId);
+                if (!page) {
+                    console.error(`Page with id "${pageId}" not found.`);
+                    return;
+                }
+
+                // Clear previous graph content
+                space.clearAll(); // Assuming a method like clearAll exists
+
+                if (hudEl) {
+                    hudEl.innerHTML = page.description;
+                }
+
+                // Create the new graph
+                if (typeof page.createGraph === 'function') {
+                    page.createGraph(space);
+                } else {
+                    console.error(`createGraph function missing for page "${pageId}"`);
+                }
+
+                // Re-initialize layout and view
+                const layoutPlugin = space.plugins.getPlugin('LayoutPlugin');
+                const layoutManager = layoutPlugin?.layoutManager;
+                if (layoutManager) {
+                    layoutManager.kick?.(); // Re-run layout physics
+                } else {
+                    console.warn('LayoutPlugin or LayoutManager not available for kicking layout.');
+                }
+
+                space.centerView(null, 0.8);
+                updateActiveButton(pageId);
+                console.log(`Loaded page: ${page.title}`);
+            }
 
             async function init() {
                 const container = S.$('#mindmap-container');
                 const contextMenuEl = S.$('#context-menu');
                 const confirmDialogEl = S.$('#confirm-dialog');
 
-                if (!container || !contextMenuEl || !confirmDialogEl) {
-                    console.error('Init Failed: Missing DOM elements.');
-                    container.innerHTML = "<p style='color:red; padding: 20px;'>Error: Missing HTML elements.</p>";
+                if (!container || !contextMenuEl || !confirmDialogEl || !pageSelectorEl || !hudEl) {
+                    console.error('Init Failed: Missing DOM elements for graph, page selector, or HUD.');
+                    if (container) container.innerHTML = "<p style='color:red; padding: 20px;'>Error: Critical HTML elements are missing.</p>";
                     return;
                 }
 
                 try {
-                    const space = new S.SpaceGraph(container, {
+                    space = new S.SpaceGraph(container, {
                         ui: {
                             contextMenuElement: contextMenuEl,
                             confirmDialogElement: confirmDialogEl,
                         },
+                        // Add any default options if needed
                     });
                     await space.init();
 
-                    createExampleGraph(space);
+                    populatePageSelector();
 
-                    const layoutPlugin = space.plugins.getPlugin('LayoutPlugin');
-                    const layoutManager = layoutPlugin?.layoutManager;
-                    const activeLayout = layoutManager?.getActiveLayout();
-
-                    if (activeLayout && layoutManager) {
-                        layoutManager.kick?.();
+                    if (pages.length > 0) {
+                        await loadPage(pages[0].id); // Load the first page by default
                     } else {
-                        console.error('LayoutPlugin or activeLayout not available.');
+                        console.warn('No pages defined in pages.js');
+                        if (hudEl) hudEl.innerHTML = "<p>No pages available to display.</p>";
                     }
 
-                    space.centerView(null, 0.8);
-                    space.animate();
+                    space.animate(); // Start animation loop
 
-                    window.space = space;
+                    window.space = space; // Expose space to global for debugging
                 } catch (error) {
                     console.error('Init Failed:', error);
                     container.innerHTML = `<div style="color: red; padding: 20px;">Error initializing: ${error.message}<br><pre>${error.stack}</pre></div>`;
