+++ src/index.css
@@ -281,6 +281,79 @@ body {
     color: var(--sg-text-on-accent);
 }
 
+/* Styles for page selector and HUD */
+#page-selector {
+    position: absolute;
+    top: 60px; /* Adjusted to be below the main toolbar */
+    left: 10px;
+    z-index: 1001; /* Above toolbar */
+    display: flex;
+    flex-direction: column; /* Stack buttons vertically */
+    gap: 5px;
+}
+
+#page-selector button {
+    padding: 8px 12px;
+    background-color: var(--sg-node-button-bg);
+    color: var(--sg-text-primary);
+    border: 1px solid var(--sg-accent-secondary);
+    border-radius: 4px;
+    cursor: pointer;
+    text-align: left;
+    transition:
+        background-color var(--sg-transition-fast),
+        border-color var(--sg-transition-fast);
+}
+
+#page-selector button:hover {
+    background-color: var(--sg-node-button-bg-hover);
+    border-color: var(--sg-accent-primary);
+}
+
+#page-selector button.active {
+    background-color: var(--sg-accent-primary);
+    color: var(--sg-text-on-accent);
+    border-color: var(--sg-accent-primary);
+}
+
+#hud {
+    position: absolute;
+    bottom: 10px;
+    left: 10px;
+    z-index: 100; /* Below page selector but above graph */
+    background-color: var(--sg-bg-ui);
+    color: var(--sg-text-primary);
+    padding: 15px;
+    border-radius: 6px;
+    border: 1px solid var(--sg-accent-secondary);
+    max-width: 350px; /* Increased max-width */
+    font-size: 0.9em;
+    box-shadow: 0 2px 10px rgba(0,0,0,0.3);
+    backdrop-filter: blur(8px);
+}
+
+#hud h3 {
+    margin-top: 0;
+    margin-bottom: 10px;
+    color: var(--sg-accent-primary);
+    font-size: 1.1em;
+}
+#hud p {
+    margin-bottom: 8px;
+    line-height: 1.4;
+}
+#hud ul {
+    margin-bottom: 8px;
+    padding-left: 1.2em;
+}
+#hud code {
+    background-color: rgba(var(--sg-accent-primary-rgb), 0.1);
+    color: var(--sg-accent-primary);
+    padding: 0.1em 0.3em;
+    border-radius: 3px;
+    font-family: var(--sg-font-mono);
+}
+
 #keyboard-shortcuts-dialog {
     min-width: 450px;
     max-width: 90vw;
